refactor(ava): rename folder
